+++ sdk/ai/ai-agents/samples/v1-beta/javascript/agentsFabric.js
@@ -8,11 +8,10 @@
  */
 
 const { AgentsClient, ToolUtility, isOutputOfType } = require("@azure/ai-agents");
-const { delay } = require("@azure/core-util");
 const { DefaultAzureCredential } = require("@azure/identity");
 require("dotenv/config");
 
-const projectEndpoint = process.env["PROJECT_ENDPOINT"] || "<project connection string>";
+const projectEndpoint = process.env["PROJECT_ENDPOINT"] || "<project endpoint>";
 const modelDeploymentName = process.env["MODEL_DEPLOYMENT_NAME"] || "gpt-4o";
 
 async function main() {
@@ -44,17 +43,17 @@ async function main() {
   );
   console.log(`Created message, message ID: ${message.id}`);
 
-  // Create and process agent run in thread with tools
-  let run = await client.runs.create(thread.id, agent.id);
-  while (run.status === "queued" || run.status === "in_progress") {
-    await delay(1000);
-    run = await client.runs.get(thread.id, run.id);
-  }
-  if (run.status === "failed") {
-    console.log(`Run failed: ${run.lastError}`);
-  }
+  // Create and poll a run
+  console.log("Creating run...");
+  const run = await client.runs.createAndPoll(thread.id, agent.id, {
+    pollingOptions: {
+      intervalInMs: 2000,
+    },
+    onResponse: (response) => {
+      console.log(`Received response with status: ${response.status}`);
+    },
+  });
   console.log(`Run finished with status: ${run.status}`);
-  console.log(`Failure: ${run.lastError?.message}`);
 
   // Delete the agent when done
   await client.deleteAgent(agent.id);
@@ -66,7 +65,7 @@ async function main() {
 
   // Get the first message
   for await (const m of messagesIterator) {
-    const agentMessage = message.content[0];
+    const agentMessage = m.content[0];
     if (isOutputOfType(agentMessage, "text")) {
       const textContent = agentMessage;
       console.log(`Text Message Content - ${textContent.text.value}`);