+++ sdk/monitor/monitor-opentelemetry-exporter/src/export/statsbeat/longIntervalStatsbeatMetrics.ts
@@ -69,9 +69,7 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
 
     this.setFeatures();
 
-    this.longIntervalStatsbeatMeterProvider = new MeterProvider();
     this.longIntervalAzureExporter = new AzureMonitorStatsbeatExporter(exporterConfig);
-
     // Export Long Interval Statsbeats every day
     const longIntervalMetricReaderOptions: PeriodicExportingMetricReaderOptions = {
       exporter: this.longIntervalAzureExporter,
@@ -82,7 +80,9 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
     this.longIntervalMetricReader = new PeriodicExportingMetricReader(
       longIntervalMetricReaderOptions,
     );
-    this.longIntervalStatsbeatMeterProvider.addMetricReader(this.longIntervalMetricReader);
+    this.longIntervalStatsbeatMeterProvider = new MeterProvider({
+      readers: [new PeriodicExportingMetricReader(longIntervalMetricReaderOptions)],
+    });
     this.longIntervalStatsbeatMeter = this.longIntervalStatsbeatMeterProvider.getMeter(
       "Azure Monitor Long Interval Statsbeat",
     );
@@ -134,7 +134,7 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
           (await this.longIntervalMetricReader.collect()).resourceMetrics,
           (result: ExportResult) => {
             if (result.code !== ExportResultCode.SUCCESS) {
-              diag.error(`LongIntervalStatsbeat: metrics export failed (error ${result.error})`);
+              diag.debug(`LongIntervalStatsbeat: metrics export failed (error ${result.error})`);
             }
           },
         );
@@ -147,7 +147,8 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
   private getEnvironmentStatus(observableResult: BatchObservableResult): void {
     this.setFeatures();
     let attributes;
-    if (this.instrumentation) {
+    // Only send instrumentation statsbeat if value is greater than zero
+    if (this.instrumentation > 0) {
       attributes = {
         ...this.commonProperties,
         feature: this.instrumentation,
@@ -156,7 +157,8 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
       observableResult.observe(this.featureStatsbeatGauge, 1, { ...attributes });
     }
 
-    if (this.feature) {
+    // Only send feature statsbeat if value is greater than zero
+    if (this.feature > 0) {
       attributes = {
         ...this.commonProperties,
         feature: this.feature,
@@ -173,7 +175,7 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
         this.feature = JSON.parse(statsbeatFeatures).feature;
         this.instrumentation = JSON.parse(statsbeatFeatures).instrumentation;
       } catch (error: any) {
-        diag.error(
+        diag.debug(
           `LongIntervalStatsbeat: Failed to parse features/instrumentations (error ${error})`,
         );
       }