+++ sdk/postgresql/arm-postgresql-flexible/src/operations/getPrivateDnsZoneSuffix.ts
@@ -6,12 +6,12 @@
  * Changes may cause incorrect behavior and will be lost if the code is regenerated.
  */
 
-import { GetPrivateDnsZoneSuffix } from "../operationsInterfaces/index.js";
+import type { GetPrivateDnsZoneSuffix } from "../operationsInterfaces/index.js";
 import * as coreClient from "@azure/core-client";
 import * as Mappers from "../models/mappers.js";
 import * as Parameters from "../models/parameters.js";
-import { PostgreSQLManagementFlexibleServerClient } from "../postgreSQLManagementFlexibleServerClient.js";
-import {
+import type { PostgreSQLManagementFlexibleServerClient } from "../postgreSQLManagementFlexibleServerClient.js";
+import type {
   GetPrivateDnsZoneSuffixExecuteOptionalParams,
   GetPrivateDnsZoneSuffixExecuteResponse,
 } from "../models/index.js";