+++ sdk/postgresql/arm-postgresql-flexible/src/operationsInterfaces/administrators.ts
@@ -6,9 +6,9 @@
  * Changes may cause incorrect behavior and will be lost if the code is regenerated.
  */
 
-import { PagedAsyncIterableIterator } from "@azure/core-paging";
-import { SimplePollerLike, OperationState } from "@azure/core-lro";
-import {
+import type { PagedAsyncIterableIterator } from "@azure/core-paging";
+import type { SimplePollerLike, OperationState } from "@azure/core-lro";
+import type {
   ActiveDirectoryAdministrator,
   AdministratorsListByServerOptionalParams,
   ActiveDirectoryAdministratorAdd,
@@ -38,7 +38,7 @@ export interface Administrators {
    * @param resourceGroupName The name of the resource group. The name is case insensitive.
    * @param serverName The name of the server.
    * @param objectId Guid of the objectId for the administrator.
-   * @param parameters The required parameters for adding an active directory administrator for a server.
+   * @param parameters The required parameters for adding an Microsoft Entra Administrator for a server.
    * @param options The options parameters.
    */
   beginCreate(
@@ -48,17 +48,14 @@ export interface Administrators {
     parameters: ActiveDirectoryAdministratorAdd,
     options?: AdministratorsCreateOptionalParams,
   ): Promise<
-    SimplePollerLike<
-      OperationState<AdministratorsCreateResponse>,
-      AdministratorsCreateResponse
-    >
+    SimplePollerLike<OperationState<AdministratorsCreateResponse>, AdministratorsCreateResponse>
   >;
   /**
    * Creates a new server.
    * @param resourceGroupName The name of the resource group. The name is case insensitive.
    * @param serverName The name of the server.
    * @param objectId Guid of the objectId for the administrator.
-   * @param parameters The required parameters for adding an active directory administrator for a server.
+   * @param parameters The required parameters for adding an Microsoft Entra Administrator for a server.
    * @param options The options parameters.
    */
   beginCreateAndWait(
@@ -69,7 +66,7 @@ export interface Administrators {
     options?: AdministratorsCreateOptionalParams,
   ): Promise<AdministratorsCreateResponse>;
   /**
-   * Deletes an Active Directory Administrator associated with the server.
+   * Deletes an Microsoft Entra Administrator associated with the server.
    * @param resourceGroupName The name of the resource group. The name is case insensitive.
    * @param serverName The name of the server.
    * @param objectId Guid of the objectId for the administrator.
@@ -82,7 +79,7 @@ export interface Administrators {
     options?: AdministratorsDeleteOptionalParams,
   ): Promise<SimplePollerLike<OperationState<void>, void>>;
   /**
-   * Deletes an Active Directory Administrator associated with the server.
+   * Deletes an Microsoft Entra Administrator associated with the server.
    * @param resourceGroupName The name of the resource group. The name is case insensitive.
    * @param serverName The name of the server.
    * @param objectId Guid of the objectId for the administrator.