+++ sdk/monitor/monitor-opentelemetry-exporter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 1.0.0-beta.32 ()
+
+### Other Changes
+
+- Removed error logging upon failure to initialize long interval statsbeat.
+- No longer send statsbeat counters when values are zero.
+
 ## 1.0.0-beta.31 (2025-04-16)
 
 ### Features Added