+++ sdk/storage/storage-common/CHANGELOG.md
@@ -1,7 +1,13 @@
-<!-- dev-tool snippets ignore -->
-
 # Release History
 
-## 1.0.0 (Unreleased)
+## 12.0.0-beta.2 (2025-05-20)
+
+### Features Added
+
+- Including all released in 12.0.0-beta.1
+
+## 12.0.0-beta.1 (2025-05-14)
+
+### Features Added
 
 - Initial Release.